fix: split collateralized adapter into two files for tslint
